+++ src/main/java/dev/checku/checkuserver/domain/subject/api/SubjectApi.java
@@ -40,9 +40,6 @@ public class SubjectApi {
     public ResponseEntity<List<GetMySubjectDto.Response>> mySubjectSearch(
             @Valid GetMySubjectDto.Request dto
     ) {
-//        String session = request.getAttribute("session").toString();
-//        mySubjectService.getMySubjects(dto, session);
-//        List<GetMySubjectDto.Response> response = mySubjectService.getMySubjects(dto, session);
 
         List<GetMySubjectDto.Response> response = mySubjectService.getMySubjects(dto);
         return ResponseEntity.ok(response);
